requirements.txt
aiohttp==3.7.3
APScheduler==3.6.3
astroid==2.4.2
async-timeout==3.0.1
asynctest==0.13.0
attrs==20.3.0
chardet==4.0.0
discord.py==1.5.1
emoji==0.6.0
idna==3.0
iniconfig==1.1.1
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==5.1.0
packaging==20.8
pluggy==0.13.1
py==1.10.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.1
pytz==2020.5
rope==0.18.0
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
tzlocal==2.1s
websockets==8.1
wrapt==1.12.1
yarl==1.6.3
